requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
pyyaml==6.0.2
psutil==6.0.0
jinja2==3.1.4
